requirements.txt
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
django-cors-headers==4.3.1
pillow==10.3.0
pytest==8.1.2
Faker==24.14.0
pytest-django==4.8.0
factory-boy==3.3.0